Moved booleans and methods for adding tiles from CalculatorView to CalculatorState.
