+++ src/test/java/view/CalculatorViewTest.java
@@ -1,5 +1,6 @@
 package view;
 
+import interface_adapter.calculator.CalculatorState;
 import org.junit.jupiter.api.Test;
 import view.CalculatorView;
 import view.component.DisplayHandComponent;
@@ -10,24 +11,25 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class CalculatorViewTest {
     private CalculatorView calculatorView;
+    private CalculatorState calculatorState
     private DisplayHandComponent displayHandComponent;
 
     @Test
     void testAddTileToDisplay() {
-        calculatorView.addClosedTile(MahjongTile.ONE_MAN);
+        calculatorState.addClosedTile(MahjongTile.ONE_MAN);
         assertEquals(1, displayHandComponent.getSelectedTiles().size());
     }
 
     @Test
     void testAddChiGroupToDisplay() {
         MahjongGroup chiiGroup = new MahjongGroup(MahjongTile.ONE_MAN, MahjongTile.TWO_MAN, MahjongTile.THREE_MAN);
-        calculatorView.addChiiGroup(chiiGroup);
+        calculatorState.addChiiGroup(chiiGroup);
         assertEquals(3, displayHandComponent.getSelectedTiles().size());
     }
 
     @Test
     void testClearDisplay() {
-        calculatorView.addClosedTile(MahjongTile.ONE_MAN);
+        calculatorState.addClosedTile(MahjongTile.ONE_MAN);
         calculatorView.getDisplayHandComponent().clearSelectedTiles();
         assertEquals(0, displayHandComponent.getSelectedTiles().size());
     }
